+++ pypdf/_writer.py
@@ -1318,6 +1318,9 @@ class PdfWriter(PdfDocCommon):
                 `use_128bit` will be ignored.
 
         """
+        if self.incremental:
+            raise NotImplementedError("Encrypting incremental PDF files is currently not supported.")
+
         if owner_password is None:
             owner_password = user_password
 
@@ -1334,7 +1337,7 @@ class PdfWriter(PdfDocCommon):
         assert self._ID
         self._encryption = Encryption.make(alg, permissions_flag, self._ID[0])
         # in case call `encrypt` again
-        entry = self._encryption.write_entry(user_password, owner_password)
+        entry = self._encryption.write_entry(user_password, owner_password, strict=self.strict)
         if self._encrypt_entry:
             # replace old encrypt_entry
             assert self._encrypt_entry.indirect_reference is not None