+++ tests/test_writer.py
@@ -1836,6 +1836,11 @@ def test_update_form_fields3(caplog, tmp_path):
         "adresa_judet": "Конференция",
     }
     writer.update_page_form_field_values(writer.pages[0], data, flatten=True)
+    # Test that we have changed the font resource from /Ubuntu to /PYPDF1
+    new_font_resource = "/PYPDF1"
+    assert new_font_resource in writer.pages[0]["/Annots"][0]["/DA"]
+    assert new_font_resource in writer.pages[0]["/Resources"]["/Font"]
+    assert new_font_resource in writer._root_object["/AcroForm"]["/DR"]["/Font"]
     writer.write(output)
     output.seek(0)
     reader = PdfReader(output)
@@ -3205,3 +3210,11 @@ def test_clone_reader_document_root__incremental__unknown_object():
     reader = PdfReader(RESOURCE_ROOT / "crazyones.pdf")
     with mock.patch.object(writer, "_collect_incremental_clone_object_ids", return_value=[*list(range(1, 23)), 42]):
         writer.clone_reader_document_root(reader)
+
+
+def test_encrypt__incremental():
+    writer = PdfWriter(RESOURCE_ROOT / "crazyones.pdf", incremental=True)
+    writer.add_blank_page(width=10, height=10)
+
+    with pytest.raises(NotImplementedError):
+        writer.encrypt(user_password="dummy")